Was able to get jQuery validate going for cptForm portion. I tried to use it for datesForm section but it's taking too much time to debug why it validates an empty string so I'm just going to drop it. I re-arragned the HTML forms names for the claim Form for this reason just as a note.
I will use regex to extract the MM/DD/YYYY portions individually and
use those to populate the site form. Seems to be the cleanest way to extract
in case the use does enter a 3-digit year or some weird number. I can't
spend too much time on data validation, it's such a pain in the azzzzz.